pyproject.toml
[project]
name = "cv-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "weasyprint>=66.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
